+++ lib/better_controller/dsl/index.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 require_relative 'turbo_stream_builder'
+require_relative 'turbo_frame_builder'
 require_relative 'response_builder'
 require_relative 'action_builder'